split usePricing
