+++ src/hooks/index.ts
@@ -2,4 +2,5 @@ export * from './useConsume'
 export * from './useMetadata'
 export * from './usePublish'
 export * from './useCompute'
-export * from './usePricing'
+export * from './useCreatePricing'
+export * from './useTrade'
