--- src/hooks/usePricing/index.ts
@@ -1,2 +1,0 @@
-export * from './usePricing'
-export * from './PriceOptions'
