+++ src/hooks/useTrade/useTrade.ts
@@ -1,16 +1,11 @@
 import { Logger } from '@oceanprotocol/lib'
 import { useState } from 'react'
 import { useOcean } from 'providers'
-import { PriceOptions } from './PriceOptions'
 import { TransactionReceipt } from 'web3-core'
 import { getBestDataTokenPrice, getFirstPool } from 'utils/dtUtils'
 import { Decimal } from 'decimal.js'
 
-interface UsePricing {
-  createPricing: (
-    dataTokenAddress: string,
-    priceOptions: PriceOptions
-  ) => Promise<TransactionReceipt | string | null>
+interface UseTrade {
   buyDT: (
     dataTokenAddress: string,
     dtAmount: number | string
@@ -19,10 +14,10 @@ interface UsePricing {
     dataTokenAddress: string,
     dtAmount: number | string
   ) => Promise<TransactionReceipt | null>
-  pricingStep?: number
-  pricingStepText?: string
-  pricingError?: string
-  pricingIsLoading: boolean
+  tradeStep?: number
+  tradeStepText?: string
+  tradeError?: string
+  tradeIsLoading: boolean
 }
 
 export const buyDTFeedback: { [key in number]: string } = {
@@ -36,106 +31,31 @@ export const sellDTFeedback: { [key in number]: string } = {
   2: '3/3 DT sold'
 }
 
-export const createPricingFeedback: { [key in number]: string } = {
-  0: '1/4 Approving DT ...',
-  1: '2/4 Approving Ocean ...',
-  2: '3/4 Creating ....',
-  3: '4/4 Pricing created'
-}
-
-function usePricing(): UsePricing {
+function useTrade(): UseTrade {
   const { ocean, status, account, accountId, config } = useOcean()
-  const [pricingIsLoading, setPricingIsLoading] = useState(false)
-  const [pricingStep, setPricingStep] = useState<number | undefined>()
-  const [pricingStepText, setPricingStepText] = useState<string | undefined>()
-  const [pricingError, setPricingError] = useState<string | undefined>()
+  const [tradeIsLoading, setTradeIsLoading] = useState(false)
+  const [tradeStep, setTradeStep] = useState<number | undefined>()
+  const [tradeStepText, setTradeStepText] = useState<string | undefined>()
+  const [tradeError, setTradeError] = useState<string | undefined>()
   const [dtSymbol, setDtSymbol] = useState<string>()
 
   function setStepBuyDT(index?: number) {
-    setPricingStep(index)
+    setTradeStep(index)
     let message
     if (index) {
       if (dtSymbol) message = buyDTFeedback[index].replace(/DT/g, dtSymbol)
       else message = buyDTFeedback[index]
-      setPricingStepText(message)
+      setTradeStepText(message)
     }
   }
   function setStepSellDT(index?: number) {
-    setPricingStep(index)
+    setTradeStep(index)
     let message
     if (index) {
       if (dtSymbol) message = sellDTFeedback[index].replace(/DT/g, dtSymbol)
       else message = sellDTFeedback[index]
-      setPricingStepText(message)
+      setTradeStepText(message)
     }
-  }
-  function setStepCreatePricing(index?: number) {
-    setPricingStep(index)
-    let message
-    if (index) {
-      if (dtSymbol)
-        message = createPricingFeedback[index].replace(/DT/g, dtSymbol)
-      else message = createPricingFeedback[index]
-      setPricingStepText(message)
-    }
-  }
-
-  async function createPricing(
-    dataTokenAddress: string,
-    priceOptions: PriceOptions
-  ): Promise<TransactionReceipt | string | null> {
-    if (!ocean || !account || !accountId) return null
-
-    let response = null
-    try {
-      setPricingIsLoading(true)
-      setPricingError(undefined)
-      setDtSymbol(await ocean.datatokens.getSymbol(dataTokenAddress))
-      setStepCreatePricing(0)
-      switch (priceOptions.type) {
-        case 'dynamic': {
-          setStepCreatePricing(2)
-          response = await ocean.pool.createDTPool(
-            accountId,
-            dataTokenAddress,
-            priceOptions.dtAmount.toString(),
-            priceOptions.weightOnDataToken,
-            priceOptions.swapFee
-          )
-          setStepCreatePricing(3)
-          return response
-        }
-        case 'fixed': {
-          if (!config.fixedRateExchangeAddress) {
-            Logger.error(`'fixedRateExchangeAddress' not set in ccnfig.`)
-            return null
-          }
-          setStepCreatePricing(2)
-          response = await ocean.fixedRateExchange.create(
-            dataTokenAddress,
-            priceOptions.price.toString(),
-            accountId
-          )
-          setStepCreatePricing(1)
-          await ocean.datatokens.approve(
-            dataTokenAddress,
-            config.fixedRateExchangeAddress,
-            String(priceOptions.dtAmount),
-            accountId
-          )
-          setStepCreatePricing(3)
-          return response
-        }
-      }
-    } catch (error) {
-      setPricingError(error.message)
-      Logger.error(error)
-    } finally {
-      setPricingStep(undefined)
-      setPricingStepText(undefined)
-      setPricingIsLoading(false)
-    }
-    return null
   }
 
   async function buyDT(
@@ -146,8 +66,8 @@ function usePricing(): UsePricing {
 
     try {
       setDtSymbol(await ocean.datatokens.getSymbol(dataTokenAddress))
-      setPricingIsLoading(true)
-      setPricingError(undefined)
+      setTradeIsLoading(true)
+      setTradeError(undefined)
       setStepBuyDT(0)
       const bestPrice = await getBestDataTokenPrice(ocean, dataTokenAddress)
       switch (bestPrice?.type) {
@@ -200,12 +120,12 @@ function usePricing(): UsePricing {
         }
       }
     } catch (error) {
-      setPricingError(error.message)
+      setTradeError(error.message)
       Logger.error(error)
     } finally {
-      setPricingStep(undefined)
-      setPricingStepText(undefined)
-      setPricingIsLoading(false)
+      setTradeStep(undefined)
+      setTradeStepText(undefined)
+      setTradeIsLoading(false)
     }
     return null
   }
@@ -221,8 +141,8 @@ function usePricing(): UsePricing {
     }
     try {
       setDtSymbol(await ocean.datatokens.getSymbol(dataTokenAddress))
-      setPricingIsLoading(true)
-      setPricingError(undefined)
+      setTradeIsLoading(true)
+      setTradeError(undefined)
       setStepSellDT(0)
       const pool = await getFirstPool(ocean, dataTokenAddress)
       if (!pool || pool.price === 0) return null
@@ -239,26 +159,25 @@ function usePricing(): UsePricing {
       Logger.log('DT sell response', sellResponse)
       return sellResponse
     } catch (error) {
-      setPricingError(error.message)
+      setTradeError(error.message)
       Logger.error(error)
     } finally {
       setStepSellDT(undefined)
-      setPricingStepText(undefined)
-      setPricingIsLoading(false)
+      setTradeStepText(undefined)
+      setTradeIsLoading(false)
     }
     return null
   }
 
   return {
-    createPricing,
     buyDT,
     sellDT,
-    pricingStep,
-    pricingStepText,
-    pricingIsLoading,
-    pricingError
+    tradeStep,
+    tradeStepText,
+    tradeIsLoading,
+    tradeError
   }
 }
 
-export { usePricing, UsePricing }
-export default usePricing
+export { useTrade, UseTrade }
+export default useTrade
